--- src/pages/faqs.tsx
@@ -1,73 +1,0 @@
-"use client";
-
-import React from "react";
-import { Typography, Card } from "@material-tailwind/react";
-
-const FAQS = [
-  {
-    title: "How do I get started?",
-    desc: "Getting started is easy! Simply [Provide a brief overview of the initial steps or link to a detailed guide].",
-  },
-  {
-    title: "Is there a free trial available?",
-    desc: "Yes, we offer a 30 days free trial so you can experience our mobile application with no commitment.",
-  },
-  {
-    title: "How can I upgrade my account?",
-    desc: "To upgrade your account, log in and navigate to the [Upgrade Account] section in your dashboard. Follow the prompts to select your preferred plan.",
-  },
-  {
-    title: "Can I cancel my subscription anytime?",
-    desc: "Absolutely, you can cancel your subscription at any time with no questions asked. Your subscription will remain active until the end of the current billing cycle.",
-  },
-  {
-    title: "How can I upgrade my account to paid?",
-    desc: "To upgrade your account, log in and navigate to the [Upgrade Account] section in your dashboard. Follow the prompts to select your preferred plan.",
-  },
-  {
-    title: "What if I need help or have technical issues?",
-    desc: "Our dedicated support team is here to assist you. Reach out via [mention preferred support channels, e.g., live chat, email, or phone], and we'll get back to you promptly.",
-  },
-];
-
-export function Faqs() {
-  return (
-    <section className="px-8 py-20">
-      <div className="container max-w-6xl mx-auto">
-        <div className="text-center">
-          <Typography variant="h1" color="blue-gray" className="mb-4"  placeholder="" onPointerEnterCapture={() => {}} onPointerLeaveCapture={() => {}}>
-            Frequently asked questions
-          </Typography>
-          <Typography
-            variant="lead"
-            className="mx-auto mb-24 !text-gray-500 lg:w-3/5"
-            placeholder="" 
-            onPointerEnterCapture={() => {}} 
-            onPointerLeaveCapture={() => {}}
-          >
-            A lot of people don&apos;t appreciate the moment until it&apos;s
-            passed. I&apos;m not trying my hardest, and I&apos;m not trying to
-            do
-          </Typography>
-        </div>
-
-        <div className="grid gap-20 md:grid-cols-1 lg:grid-cols-3">
-          {FAQS.map(({ title, desc }) => (
-            <Card key={title} shadow={false} color="transparent"  placeholder="" onPointerEnterCapture={() => {}} onPointerLeaveCapture={() => {}}>
-              <Typography color="blue-gray" className="pb-6" variant="h4"  placeholder="" onPointerEnterCapture={() => {}} onPointerLeaveCapture={() => {}}>
-                {title}
-              </Typography>
-              <div className="pt-2">
-                <Typography className="font-normal !text-gray-500"  placeholder="" onPointerEnterCapture={() => {}} onPointerLeaveCapture={() => {}}>
-                  {desc}
-                </Typography>
-              </div>
-            </Card>
-          ))}
-        </div>
-      </div>
-    </section>
-  );
-}
-
-export default Faqs;
