+++ src/pages/feature.tsx
@@ -35,8 +35,29 @@ const ApiProviders = () => {
   return (
     <div className="flex flex-col items-center py-10">
       <h2 className="text-3xl font-bold mb-6">Simplifying the integration of APIs into Web3 projects</h2>
-      <div className="grid grid-cols-1 md:grid-cols-5 gap-8">
-        {providers.map((provider, index) => (
+      {/* <div className="grid grid-cols-1 md:grid-cols-5 gap-8"> */}
+
+
+      <div className="flex flex-col md:flex-row justify-center items-center gap-8">
+          <div className="flex flex-col items-center border border-gray-600 rounded-lg py-6 px-12">
+            <span className="text-3xl font-bold text-green-400">1000+</span>
+            <span className="text-lg text-gray-400">users per month</span>
+          </div>
+          <div className="flex flex-col items-center bg-gradient-to-t from-purple-500 to-green-500 rounded-lg py-6 px-12 mb-12">
+            <span className="text-3xl font-bold text-white">92%</span>
+            <span className="text-lg text-gray-200">
+              of users are satisfied with the platform
+            </span>
+          </div>
+          <div className="flex flex-col items-center border border-gray-600 rounded-lg py-6 px-12">
+            <span className="text-3xl font-bold text-green-400">5000+</span>
+            <span className="text-lg text-gray-400">waiting hours reduced</span>
+          </div>
+        {/* </div> */}
+
+
+
+        {/* {providers.map((provider, index) => (
           <div
             key={index}
             className="bg-green-500 hover:bg-green-900 p-6 rounded-lg text-center transition-colors duration-300 ease-in-out"
@@ -46,7 +67,7 @@ const ApiProviders = () => {
             </div>
             <p className="text-lg text-white font-bold">+{provider.name}</p>
           </div>
-        ))}
+        ))} */}
       </div>
     </div>
   );
