+++ src/pages/signup-form.tsx
@@ -7,6 +7,42 @@ import ConnectWalletModal from './connect-wallet'; // Import the modal
 import { FixedPlugin } from "@/components";
 
 
+// const SignUpForm: React.FC = () => {
+//   const [step, setStep] = useState(1);
+//   const [name, setName] = useState('');
+//   const [email, setEmail] = useState('');
+//   const [password, setPassword] = useState('');
+//   const [confirmPassword, setConfirmPassword] = useState('');
+//   const [isButtonActive, setIsButtonActive] = useState(false);
+
+//   const [isModalOpen, setModalOpen] = useState(false);
+
+//   const openModal = () => setModalOpen(true);
+//   const closeModal = () => setModalOpen(false);
+
+//   const handleNextStep = () => {
+//     if (step === 1) {
+//       if (name && email.includes('@')) {
+//         setStep(2);
+//         setIsButtonActive(false); // Reset button for next step
+//       }
+//     } else if (step === 2) {
+//       if (password && password === confirmPassword) {
+//         setStep(3);
+//         setIsButtonActive(true);
+//       }
+//     }
+//   };
+
+//   const handleInputChange = () => {
+//     if (step === 1) {
+//       setIsButtonActive(!!(name && email.includes('@')));
+//     } else if (step === 2) {
+//       setIsButtonActive(!!(password && confirmPassword));
+//     }
+//   };
+
+
 const SignUpForm: React.FC = () => {
   const [step, setStep] = useState(1);
   const [name, setName] = useState('');
@@ -14,13 +50,14 @@ const SignUpForm: React.FC = () => {
   const [password, setPassword] = useState('');
   const [confirmPassword, setConfirmPassword] = useState('');
   const [isButtonActive, setIsButtonActive] = useState(false);
-
   const [isModalOpen, setModalOpen] = useState(false);
+  const [error, setError] = useState('');
+  const [successMessage, setSuccessMessage] = useState('');
 
   const openModal = () => setModalOpen(true);
   const closeModal = () => setModalOpen(false);
 
-  const handleNextStep = () => {
+  const handleNextStep = async() => {
     if (step === 1) {
       if (name && email.includes('@')) {
         setStep(2);
@@ -28,8 +65,8 @@ const SignUpForm: React.FC = () => {
       }
     } else if (step === 2) {
       if (password && password === confirmPassword) {
-        setStep(3);
-        setIsButtonActive(true);
+        await handleSubmit(); // Call the API when user reaches the last step
+        setStep(3)
       }
     }
   };
@@ -40,6 +77,43 @@ const SignUpForm: React.FC = () => {
     } else if (step === 2) {
       setIsButtonActive(!!(password && confirmPassword));
     }
+  };
+
+  const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
+    if (e) e.preventDefault();
+
+    try {
+      // Clear any previous error messages
+      setError('');
+      setSuccessMessage('');
+
+      // Make the POST request
+      const response = await fetch('https://morlabsprotocol.vercel.app/waitlist', {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify({ email }),
+      });
+
+      if (!response.ok) {
+        throw new Error('Network response was not ok');
+    }
+
+    const result = await response.json();
+    console.log('Successfully joined the waitlist', result);
+
+  } catch (err) {
+    setError('There was a problem with the fetch operation:', err);
+  }
+    //     setSuccessMessage('You have successfully joined the waitlist!');
+    //     setStep(2); // Move to the final step after successful submission
+    //   } else {
+    //     setError('Failed to join the waitlist. Please try again.');
+    //   }
+    // } catch (err) {
+    //   setError('An error occurred. Please try again later.');
+    // }
   };
 
 // const SignUpForm: React.FC = () => {
@@ -79,6 +153,93 @@ const SignUpForm: React.FC = () => {
       </div>
 
       {step === 1 && (
+        <>
+          <form onSubmit={handleSubmit} className="space-y-4">
+            <input
+              type="text"
+              placeholder="Tell us your name..."
+              value={name}
+              onChange={(e) => {
+                setName(e.target.value);
+                handleInputChange();
+              }}
+              className="w-full px-4 py-2 border border-gray-300 rounded-lg focus:outline-none"
+            />
+            <input
+              type="email"
+              placeholder="Enter your email address..."
+              value={email}
+              onChange={(e) => {
+                setEmail(e.target.value);
+                handleInputChange();
+              }}
+              className="w-full px-4 py-2 border border-gray-300 rounded-lg focus:outline-none"
+            />
+            <button 
+        type="submit" 
+        className={`w-full px-4 py-2 mt-4 rounded-lg ${
+          isButtonActive ? 'bg-green-500 text-white' : 'bg-gray-400 text-gray-300 cursor-not-allowed'
+        }`}
+        disabled={!isButtonActive}
+      >
+        Continue
+      </button>
+          </form>
+        </>
+      )}
+
+      {step === 2 && (
+        <>
+          <form onSubmit={handleSubmit} className="space-y-4">
+            <input
+              type="password"
+              placeholder="Enter password..."
+              value={password}
+              onChange={(e) => {
+                setPassword(e.target.value);
+                handleInputChange();
+              }}
+              className="w-full px-4 py-2 border border-gray-300 rounded-lg focus:outline-none"
+            />
+            <input
+              type="password"
+              placeholder="Confirm password..."
+              value={confirmPassword}
+              onChange={(e) => {
+                setConfirmPassword(e.target.value);
+                handleInputChange();
+              }}
+              className="w-full px-4 py-2 border border-gray-300 rounded-lg focus:outline-none"
+            />
+             {/* <button 
+        type="submit" 
+        className={`w-full px-4 py-2 mt-4 rounded-lg ${
+          isButtonActive ? 'bg-green-500 text-white' : 'bg-gray-400 text-gray-300 cursor-not-allowed'
+        }`}
+        disabled={!isButtonActive}
+      >
+        Continue
+      </button> */}
+          </form>
+        </>
+      )}
+
+      {error && <p className="text-red-500 mt-4">{error}</p>}
+      {successMessage && <p className="text-green-500 mt-4">{successMessage}</p>}
+
+      {/* <button
+        onClick={handleNextStep}
+        disabled={!isButtonActive}
+        className={`mt-4 px-4 py-2 rounded-lg ${isButtonActive ? 'bg-green-500' : 'bg-gray-300'}`}
+      >
+        {step === 1 ? 'Next' : 'Join Waitlist'}
+      </button> */}
+
+
+
+
+
+      {/* {step === 1 && (
           <>
             <form className="space-y-4">
               <input
@@ -130,7 +291,7 @@ const SignUpForm: React.FC = () => {
               />
             </form>
           </>
-        )}
+        )} */}
 
         {step === 3 && (
           <div className="text-center">
@@ -148,7 +309,7 @@ const SignUpForm: React.FC = () => {
           </div>
         )}
 
-        {step < 3 && (
+        {step > 1 && (
           <button
             className={`w-full px-4 py-2 mt-4 rounded-lg ${
               isButtonActive ? 'bg-green-500 text-white' : 'bg-gray-400 text-gray-300 cursor-not-allowed'
